requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
asyncpg==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1